+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.1.1] - 2026-04-22
+### Fixed
+- Install powershell for AUR build
+- Invoke global tools directly
+
 ## [7.1.0] - 2026-04-15
 ### Changed
 - Updated dependencies
@@ -477,6 +482,7 @@ have been added:
 ### Added
 - Initial release
 
+[7.1.1]: https://github.com/Skycoder42/dart_test_tools/compare/v7.1.0...v7.1.1
 [7.1.0]: https://github.com/Skycoder42/dart_test_tools/compare/v7.0.3...v7.1.0
 [7.0.3]: https://github.com/Skycoder42/dart_test_tools/compare/v7.0.2+1...v7.0.3
 [7.0.2+1]: https://github.com/Skycoder42/dart_test_tools/compare/v7.0.2...v7.0.2+1